Receiving data correctly to edit item

+++ catina/src/models/Item.java
@@ -83,7 +83,7 @@ public class Item extends ModelStandart{
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
